+++ ic-management-canister-types/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Used `candid:Reserved` inside `TakenFromCanister` and `MetadataUpload` variants in `SnapshotSource`.
 - Renamed `MainMemory` to `WasmMemory` in `SnapshotDataKind` and `SnapshotDataOffset`.
+- Added `source` field to `LoadSnapshotRecord`.
 
 While this is technically a breaking change in the Rust type system, we are treating it as a patch fix.
 This is because the affected types and methods are for new, unreleased features (snapshot download/upload).